+++ src/main/kotlin/com/github/victorrentea/livecoding/varie/DontOverrideBeforeInspection.kt
@@ -5,17 +5,13 @@ import com.intellij.psi.PsiMethod
 import com.siyeh.ig.BaseInspection
 import com.siyeh.ig.BaseInspectionVisitor
 
-
 class DontOverrideBeforeInspection : BaseInspection() {
+    override fun buildErrorString(vararg infos: Any?) =
+        "Dangerous override of inherited before method."
 
     override fun buildVisitor() = DontOverrideBeforeVisitor()
 
-    override fun buildErrorString(vararg infos: Any?): String {
-        return "Dangerous override of inherited before method."
-    }
-
     class DontOverrideBeforeVisitor : BaseInspectionVisitor() {
-
         override fun visitMethod(method: PsiMethod?) {
             if (method == null) return